requirements.txt
Flask>=2.0.0
Werkzeug>=2.0.0
Biopython>=1.79
requests>=2.25.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
numpy>=1.20.0
scikit-learn>=1.0.0
joblib>=1.1.0
umap-learn>=0.5.0
plotly>=5.10.0
kaleido>=0.2.1
gunicorn>=20.1.0
shap>=0.40.0
